FDG-11430 error boundary implementation for savings bonds
