+++ src/layouts/explainer/sections/treasury-savings-bonds/purchase-of-savings-bonds/savings-bonds-sold-by-type-chart/savings-bonds-sold-by-type-chart.tsx
@@ -3,7 +3,15 @@ import ChartContainer from '../../../../explainer-components/chart-container/cha
 import { chartStyle, loadingIcon, container } from './savings-bonds-sold-by-type-chart.module.scss';
 import { Area, AreaChart, CartesianGrid, ReferenceLine, ResponsiveContainer, Tooltip, XAxis, YAxis } from 'recharts';
 import ChartLegend from './chart-legend/chart-legend';
-import { chartCopy, getXAxisValues, savingsBonds, savingsBondsMap, yAxisFormatter } from './savings-bonds-sold-by-type-chart-helper';
+import {
+  chartCopy,
+  fyEndpoint,
+  getXAxisValues,
+  savingsBonds,
+  savingsBondsMap,
+  sortByType,
+  yAxisFormatter,
+} from './savings-bonds-sold-by-type-chart-helper';
 import CustomTooltip from './custom-tooltip/custom-tooltip';
 import ChartHeader from './chart-header/chart-header';
 import ChartDescription from './chart-description/chart-description';
@@ -11,6 +19,16 @@ import { analyticsEventHandler } from '../../../../explainer-helpers/explainer-h
 import { ga4DataLayerPush } from '../../../../../../helpers/google-analytics/google-analytics-helper';
 import globalConstants from '../../../../../../helpers/constants';
 import LoadingIndicator from '../../../../../../components/loading-indicator/loading-indicator';
+import { treasurySavingsBondsExplainerSecondary } from '../../treasury-savings-bonds.module.scss';
+import { getShortForm } from '../../../../../../utils/rounding-utils';
+import VisualizationCallout from '../../../../../../components/visualization-callout/visualization-callout';
+import { graphql, useStaticQuery } from 'gatsby';
+import { getSaleBondsFootNotes } from '../../learn-more/learn-more-helper';
+import { apiPrefix, basicFetch } from '../../../../../../utils/api-utils';
+import { adjustDataForInflation } from '../../../../../../helpers/inflation-adjust/inflation-adjust';
+import { getDateWithoutTimeZoneAdjust } from '../../../../../../utils/date-utils';
+import { ICpiDataMap } from '../../../../../../models/ICpiDataMap';
+import { useErrorBoundary } from 'react-error-boundary';
 
 export interface ISavingBondsByTypeChartData {
   year: string;
@@ -26,25 +44,115 @@ export interface ISavingBondsByTypeChartData {
   K?: number;
 }
 
-interface ISavingsBondsSoldByTypeChart {
-  chartData: ISavingBondsByTypeChartData[];
-  inflationChartData: ISavingBondsByTypeChartData[];
-  curFy: number;
-  chartDate: Date;
-}
-
 let gaTimer;
 
-const SavingsBondsSoldByTypeChart: FunctionComponent<ISavingsBondsSoldByTypeChart> = ({ chartData, inflationChartData, curFy, chartDate }) => {
+interface BondSaleEntry {
+  year: string;
+  [key: string]: string;
+}
+
+type SalesData = Record<string, number>;
+
+const SavingsBondsSoldByTypeChart: FunctionComponent = ({ cpiDataByYear }) => {
   const [selectedChartView, setSelectedChartView] = useState<string>('amounts');
   const [hiddenFields, setHiddenFields] = useState<string[]>([]);
-  const chartTitle = `Savings Bonds Sold by Type Over Time, FY 1935 – FYTD ${curFy ?? '--'}`;
   const [sortedBonds, setSortedBonds] = useState<string[]>();
   const [maxYear, setMaxYear] = useState<number>();
   const [xAxis, setXAxis] = useState<number[]>();
   const [inflationSwitch, setInflationSwitch] = useState<boolean>(false);
   const [chartFocus, setChartFocus] = useState<boolean>(false);
   const [chartHover, setChartHover] = useState<boolean>(false);
+  const [chartData, setChartData] = useState<ISavingBondsByTypeChartData[]>();
+  const [curFy, setCurFy] = useState<string>();
+  const [historyChartDate, setHistoryChartDate] = useState<Date>(null);
+  const [inflationChartData, setInflationChartData] = useState<ISavingBondsByTypeChartData[]>();
+  const [mostBondSalesYear, setMostBondSalesYear] = useState<string | null>(null);
+  const [mostBondSales, setMostBondSales] = useState<number>(0);
+  const [secondMostBondSalesYear, setSecondMostBondSalesYear] = useState<string | null>(null);
+  const [secondMostBondSales, setSecondMostBondSales] = useState<number>(0);
+  const chartTitle = `Savings Bonds Sold by Type Over Time, FY 1935 – FYTD ${curFy ?? '--'}`;
+
+  const { showBoundary } = useErrorBoundary();
+
+  const allSavingsBondsByTypeHistorical = useStaticQuery(
+    graphql`
+      query {
+        allSavingsBondsByTypeHistoricalCsv {
+          savingsBondsByTypeHistoricalCsv: nodes {
+            year
+            bond_type
+            sales
+          }
+        }
+      }
+    `
+  );
+  let savingsBondsByTypeHistorical = allSavingsBondsByTypeHistorical.allSavingsBondsByTypeHistoricalCsv.savingsBondsByTypeHistoricalCsv;
+
+  const savingsBondsEndpoint = 'v1/accounting/od/securities_sales?filter=security_type_desc:eq:Savings%20Bond';
+
+  useEffect(() => {
+    basicFetch(`${apiPrefix}${savingsBondsEndpoint}&page[size]=1`).then(metaRes => {
+      if (metaRes.meta && typeof metaRes.meta['total-pages'] !== 'undefined') {
+        const pageSize = metaRes.meta['total-pages'];
+        basicFetch(`${apiPrefix}ww${savingsBondsEndpoint}&page[size]=${pageSize}`)
+          .then(res => {
+            if (res.data) {
+              const currentData = sortByType(res.data, 'record_fiscal_year', 'security_class_desc', 'net_sales_amt');
+              const historicalData = sortByType(savingsBondsByTypeHistorical, 'year', 'bond_type', 'sales');
+              const allData = [...historicalData, ...currentData].sort((a, b) => a.year - b.year);
+
+              res.data = adjustDataForInflation(res.data, 'net_sales_amt', 'record_fiscal_year', cpiDataByYear);
+              const inflationCurrentData = sortByType(res.data, 'record_fiscal_year', 'security_class_desc', 'net_sales_amt');
+
+              savingsBondsByTypeHistorical = adjustDataForInflation(savingsBondsByTypeHistorical, 'sales', 'year', cpiDataByYear);
+              const inflationHistoricalData = sortByType(savingsBondsByTypeHistorical, 'year', 'bond_type', 'sales');
+              const inflationAllData = [...inflationHistoricalData, ...inflationCurrentData]
+                .sort((a, b) => a.year - b.year)
+                .filter(entry => cpiDataByYear[entry.year]);
+
+              const salesByYear: SalesData = allData.reduce((acc, entry: BondSaleEntry) => {
+                const totalSalesForYear = acc[entry.year] || 0;
+                const yearlySales = Object.keys(entry)
+                  .filter(key => key !== 'year')
+                  .reduce((sum, key) => sum + Number(entry[key]), 0);
+
+                acc[entry.year] = totalSalesForYear + yearlySales;
+                return acc;
+              }, {});
+
+              const sortedYears = Object.entries(salesByYear)
+                .map(([year, totalSales]) => ({ year, totalSales }))
+                .sort((a, b) => b.totalSales - a.totalSales);
+
+              if (sortedYears.length > 0) {
+                setMostBondSalesYear(sortedYears[0].year);
+                setMostBondSales(sortedYears[0].totalSales);
+                if (sortedYears.length > 1) {
+                  setSecondMostBondSalesYear(sortedYears[1].year);
+                  setSecondMostBondSales(sortedYears[1].totalSales);
+                }
+              }
+              setChartData(allData);
+              setInflationChartData(inflationAllData);
+            }
+          })
+          .catch(err => {
+            showBoundary(err);
+          });
+      }
+    });
+  }, []);
+
+  useEffect(() => {
+    basicFetch(`${apiPrefix}${fyEndpoint}`).then(res => {
+      if (res.data) {
+        const data = res.data[0];
+        setCurFy(data.record_fiscal_year);
+        setHistoryChartDate(getDateWithoutTimeZoneAdjust(data.record_date));
+      }
+    });
+  }, []);
 
   let activeChartData = inflationSwitch ? inflationChartData : chartData;
   const handleInflationToggle = (isAdjusted: boolean) => {
@@ -110,7 +218,7 @@ const SavingsBondsSoldByTypeChart: FunctionComponent<ISavingsBondsSoldByTypeChar
   return (
     <>
       <div className={container}>
-        <ChartContainer title={chartTitle} altText={chartCopy.altText} date={chartDate} footer={chartCopy.footer} header={header}>
+        <ChartContainer title={chartTitle} altText={chartCopy.altText} date={historyChartDate} footer={chartCopy.footer} header={header}>
           {selectedChartView === 'amounts' &&
             (!chartData || !sortedBonds ? (
               <LoadingIndicator loadingClass={loadingIcon} />
@@ -165,6 +273,13 @@ const SavingsBondsSoldByTypeChart: FunctionComponent<ISavingsBondsSoldByTypeChar
           {selectedChartView === 'description' && <ChartDescription />}
         </ChartContainer>
       </div>
+      <VisualizationCallout color={treasurySavingsBondsExplainerSecondary}>
+        <p>
+          Savings bonds were most popular in {mostBondSalesYear ?? '--'} and {secondMostBondSalesYear ?? '--'} when $
+          {mostBondSales ? getShortForm(mostBondSales) : '--'} and ${secondMostBondSales ? getShortForm(secondMostBondSales) : '--'} bonds were sold,
+          respectively.
+        </p>
+      </VisualizationCallout>
     </>
   );
 };
