+++ src/layouts/explainer/sections/treasury-savings-bonds/purchase-of-savings-bonds/what-influences-purchase-of-savings-bonds.tsx
@@ -1,5 +1,5 @@
-import React, { FunctionComponent, useEffect, useState } from 'react';
-import SavingsBondsSoldByTypeChart, { ISavingBondsByTypeChartData } from './savings-bonds-sold-by-type-chart/savings-bonds-sold-by-type-chart';
+import React, { FunctionComponent } from 'react';
+import SavingsBondsSoldByTypeChart from './savings-bonds-sold-by-type-chart/savings-bonds-sold-by-type-chart';
 import VisualizationCallout from '../../../../../components/visualization-callout/visualization-callout';
 import { visWithCallout } from '../../../explainer.module.scss';
 import { treasurySavingsBondsExplainerSecondary } from '../treasury-savings-bonds.module.scss';
@@ -7,120 +7,21 @@ import { subsectionHeader } from './what-influences-purchase-of-savings-bonds.mo
 import ImageContainer from '../../../explainer-components/image-container/image-container';
 import BondPoster from '../../../../../../static/images/savings-bonds/Bond-Poster.png';
 import PresidentKennedy from '../../../../../../static/images/savings-bonds/President-Kennedy-Holding-Bond.png';
-import { apiPrefix, basicFetch } from '../../../../../utils/api-utils';
-import { getShortForm } from '../../../../../utils/rounding-utils';
-import IBondSalesChart from './i-bond-sales-chart/i-bond-sales-chart';
-import { graphql, useStaticQuery } from 'gatsby';
-import { fyEndpoint, sortByType } from './savings-bonds-sold-by-type-chart/savings-bonds-sold-by-type-chart-helper';
-import { getDateWithoutTimeZoneAdjust } from '../../../../../utils/date-utils';
 import AnchorText from '../../../../../components/anchor-text/anchor-text';
 import { getSaleBondsFootNotes } from '../learn-more/learn-more-helper';
-import { adjustDataForInflation } from '../../../../../helpers/inflation-adjust/inflation-adjust';
 import { ICpiDataMap } from '../../../../../models/ICpiDataMap';
 import { analyticsEventHandler } from '../../../explainer-helpers/explainer-helpers';
-import { ErrorBoundary, useErrorBoundary } from 'react-error-boundary';
+import IBondSalesChart from './i-bond-sales-chart/i-bond-sales-chart';
+import { ErrorBoundary } from 'react-error-boundary';
 import ChartApiError from '../../../explainer-components/chart-api-error/chart-api-error';
 
-interface BondSaleEntry {
-  year: string;
-  [key: string]: string;
-}
-
-type SalesData = Record<string, number>;
 type CalloutProps = {
   cpiDataByYear: ICpiDataMap[];
   cpi12MonthPercentChange: ICpiDataMap[];
 };
 
 const WhatInfluencesPurchaseOfSavingsBonds: FunctionComponent = ({ cpi12MonthPercentChange, cpiDataByYear }: CalloutProps) => {
-  const [chartData, setChartData] = useState<ISavingBondsByTypeChartData[]>();
-  const [curFy, setCurFy] = useState<string>();
-  const [historyChartDate, setHistoryChartDate] = useState<Date>(null);
-  const [inflationChartData, setInflationChartData] = useState<ISavingBondsByTypeChartData[]>();
-  const [mostBondSalesYear, setMostBondSalesYear] = useState<string | null>(null);
-  const [mostBondSales, setMostBondSales] = useState<number>(0);
-  const [secondMostBondSalesYear, setSecondMostBondSalesYear] = useState<string | null>(null);
-  const [secondMostBondSales, setSecondMostBondSales] = useState<number>(0);
-  const { showBoundary } = useErrorBoundary();
-  const allSavingsBondsByTypeHistorical = useStaticQuery(
-    graphql`
-      query {
-        allSavingsBondsByTypeHistoricalCsv {
-          savingsBondsByTypeHistoricalCsv: nodes {
-            year
-            bond_type
-            sales
-          }
-        }
-      }
-    `
-  );
-  let savingsBondsByTypeHistorical = allSavingsBondsByTypeHistorical.allSavingsBondsByTypeHistoricalCsv.savingsBondsByTypeHistoricalCsv;
-
-  const savingsBondsEndpoint = 'v1/accounting/od/securities_sales?filter=security_type_desc:eq:Savings%20Bond';
   const anchor = getSaleBondsFootNotes()[1];
-  useEffect(() => {
-    basicFetch(`${apiPrefix}${savingsBondsEndpoint}&page[size]=1`).then(metaRes => {
-      if (metaRes.meta && typeof metaRes.meta['total-pages'] !== 'undefined') {
-        const pageSize = metaRes.meta['total-pages'];
-        basicFetch(`${apiPrefix}${savingsBondsEndpoint}&page[size]=${pageSize}`)
-          .then(res => {
-            if (res.data) {
-              const currentData = sortByType(res.data, 'record_fiscal_year', 'security_class_desc', 'net_sales_amt');
-              const historicalData = sortByType(savingsBondsByTypeHistorical, 'year', 'bond_type', 'sales');
-              const allData = [...historicalData, ...currentData].sort((a, b) => a.year - b.year);
-
-              res.data = adjustDataForInflation(res.data, 'net_sales_amt', 'record_fiscal_year', cpiDataByYear);
-              const inflationCurrentData = sortByType(res.data, 'record_fiscal_year', 'security_class_desc', 'net_sales_amt');
-
-              savingsBondsByTypeHistorical = adjustDataForInflation(savingsBondsByTypeHistorical, 'sales', 'year', cpiDataByYear);
-              const inflationHistoricalData = sortByType(savingsBondsByTypeHistorical, 'year', 'bond_type', 'sales');
-              const inflationAllData = [...inflationHistoricalData, ...inflationCurrentData]
-                .sort((a, b) => a.year - b.year)
-                .filter(entry => cpiDataByYear[entry.year]);
-
-              const salesByYear: SalesData = allData.reduce((acc, entry: BondSaleEntry) => {
-                const totalSalesForYear = acc[entry.year] || 0;
-                const yearlySales = Object.keys(entry)
-                  .filter(key => key !== 'year')
-                  .reduce((sum, key) => sum + Number(entry[key]), 0);
-
-                acc[entry.year] = totalSalesForYear + yearlySales;
-                return acc;
-              }, {});
-
-              const sortedYears = Object.entries(salesByYear)
-                .map(([year, totalSales]) => ({ year, totalSales }))
-                .sort((a, b) => b.totalSales - a.totalSales);
-
-              if (sortedYears.length > 0) {
-                setMostBondSalesYear(sortedYears[0].year);
-                setMostBondSales(sortedYears[0].totalSales);
-                if (sortedYears.length > 1) {
-                  setSecondMostBondSalesYear(sortedYears[1].year);
-                  setSecondMostBondSales(sortedYears[1].totalSales);
-                }
-              }
-              setChartData(allData);
-              setInflationChartData(inflationAllData);
-            }
-          })
-          .catch(err => {
-            showBoundary(err);
-          });
-      }
-    });
-  }, []);
-
-  useEffect(() => {
-    basicFetch(`${apiPrefix}${fyEndpoint}`).then(res => {
-      if (res.data) {
-        const data = res.data[0];
-        setCurFy(data.record_fiscal_year);
-        setHistoryChartDate(getDateWithoutTimeZoneAdjust(data.record_date));
-      }
-    });
-  }, []);
 
   const footnoteClick = () => {
     analyticsEventHandler('Savings Bonds - Footnote Click', 'Footnote Click');
@@ -162,14 +63,9 @@ const WhatInfluencesPurchaseOfSavingsBonds: FunctionComponent = ({ cpi12MonthPer
       </ImageContainer>
       <p>The chart below shows savings bond sales over time for all savings bond types.</p>
       <figure className={visWithCallout}>
-        <SavingsBondsSoldByTypeChart chartData={chartData} curFy={curFy} chartDate={historyChartDate} inflationChartData={inflationChartData} />
-        <VisualizationCallout color={treasurySavingsBondsExplainerSecondary}>
-          <p>
-            Savings bonds were most popular in {mostBondSalesYear ?? '--'} and {secondMostBondSalesYear ?? '--'} when $
-            {mostBondSales ? getShortForm(mostBondSales) : '--'} and ${secondMostBondSales ? getShortForm(secondMostBondSales) : '--'} bonds were
-            sold, respectively.
-          </p>
-        </VisualizationCallout>
+        <ErrorBoundary fallback={<ChartApiError />}>
+          <SavingsBondsSoldByTypeChart cpiDataByYear={cpiDataByYear} />
+        </ErrorBoundary>
       </figure>
       <h5 className={subsectionHeader}>Interest Rates and Inflation</h5>
       <p>
@@ -185,7 +81,7 @@ const WhatInfluencesPurchaseOfSavingsBonds: FunctionComponent = ({ cpi12MonthPer
         bonds between April 2021 and February 2023. The chart below shows inflation data and I bond purchases from the last 15 years.
       </p>
       <figure className={visWithCallout}>
-        <IBondSalesChart cpi12MonthPercentChange={cpi12MonthPercentChange} curFy={curFy} />
+        <IBondSalesChart cpi12MonthPercentChange={cpi12MonthPercentChange} />
         <VisualizationCallout color={treasurySavingsBondsExplainerSecondary}>
           <p>Generally, higher inflation rates are correlated with an increase in demand for inflation-protected securities like I bonds.</p>
         </VisualizationCallout>
